Auth config & builder draft
